requirements.txt
Flask==1.1.2
matplotlib==3.2.2
Keras==2.4.3 
Keras-Preprocessing==1.1.2
numpy==1.18.5
requests==2.23.0
h5py==2.10.0   
pandas==1.1.2 
opencv-python==4.1.2.30
scikit-learn==0.22.2.post1         
scipy==1.4.1 
tensorflow==2.3.0 
Pillow==7.0.0
flask-ngrok==0.0.25